crawl-service/requirements.txt
httpx>=0.27.0
selectolax>=0.3.21
beautifulsoup4>=4.12.0
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.10.0
apscheduler>=3.10.0
mysql-connector-python>=9.0.0
python-dotenv>=1.0.0
